requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
django-environ==0.12.0
djangorestframework==3.16.0
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
python-dotenv==1.1.1
requests==2.32.4
ruff==0.12.7
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.9.0
